% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPA.R
\name{GNPA}
\alias{GNPA}
\title{Générateur de nombres pseudo-aléatoire}
\usage{
GNPA(n, a = 41358, m = 2147483647, x0)
}
\arguments{
\item{n}{nombre d'uniformes}

\item{a}{multiplicateur (41358)}

\item{m}{modulo (2^31 - 1)}

\item{x0}{valeur de départ}
}
\description{
Les fonctions ne sont pas idiotproof, il faut rentrer les paramètres correctement.
}

