% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rente discrète qx.R
\name{ren_dis_qx}
\alias{ren_dis_qx}
\title{Rente discrète avec vecteur qx}
\usage{
ren_dis_qx(x = 0, n = 1, m = 0, g = 1, delta = 0, qx, s = 0,
  kappa = 0)
}
\arguments{
\item{x}{âge de l'individu (0 par défaut)}

\item{n}{durée de contrat (1)}

\item{m}{années différée(0)}

\item{g}{paiement de rente (1)}

\item{delta}{taux d'intérêt (0)}

\item{qx}{vecteur des qx}

\item{s}{valeurs pour évaluer Fz (0)}

\item{kappa}{kappa comme d'habitude (0)}
}
\description{
Il faut seulement entrer les paramètres et tout ce qu'Étienne demande normalement est calculé.
Les fonctions ne sont pas idiotproof, il faut rentrer les paramètres correctement.
}

